Move font-loader component to src/component
